refactor: extract rest runner from run.go and clean up (LAB-961)
Extract REST-specific runner logic (Config, runTest, template loading,
filtering) from run.go into rest.go. Remove templates.go (merged into
rest.go). Clean up README docs and remove stale env var unsets in LLM
tests.

This fixes the CI build failure caused by the committed run.go still
importing pkg/owasp after the directory was renamed to pkg/orchestrator.
